backend/requirements.txt
# CMC IP Marketplace - Backend Dependencies

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Supabase
supabase==2.10.0

# Anthropic
anthropic==0.44.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.10.0
pydantic-settings==2.7.0
email-validator==2.2.0

# File handling
pypdf==6.7.1
python-magic==0.4.27
pillow==11.0.0

# Utilities
httpx==0.27.0

# Monitoring
sentry-sdk[fastapi]==2.53.0
